Creating a measurement class
