+++ OHIFViewer/client/components/viewer/viewer.js
@@ -46,8 +46,6 @@ Meteor.startup(() => {
     Session.setDefault('activeViewport', false);
     Session.setDefault('leftSidebar', false);
     Session.setDefault('rightSidebar', false);
-    Session.set('TimepointsReady', false);
-    Session.set('MeasurementsReady', false);
 
     OHIF.viewer.defaultTool = 'wwwc';
     OHIF.viewer.refLinesEnabled = true;
@@ -76,6 +74,9 @@ Meteor.startup(() => {
     // Metadata configuration
     const metadataProvider = OHIF.viewer.metadataProvider;
     cornerstone.metaData.addProvider(metadataProvider.provider.bind(metadataProvider));
+
+    // Instanciate viewer plugins
+    OHIF.viewer.measurementTable = new OHIF.measurementTable();
 });
 
 Template.viewer.onCreated(() => {
@@ -85,20 +86,6 @@ Template.viewer.onCreated(() => {
 
     // Define the OHIF.viewer.data global object
     OHIF.viewer.data = OHIF.viewer.data || Session.get('ViewerData') || {};
-
-    const { TimepointApi, MeasurementApi } = OHIF.measurements;
-
-    OHIF.viewer.data.currentTimepointId = 'TimepointId';
-
-    const timepointApi = new TimepointApi(OHIF.viewer.data.currentTimepointId);
-    const measurementApi = new MeasurementApi(timepointApi);
-    const apis = {
-        timepointApi,
-        measurementApi
-    };
-
-    Object.assign(OHIF.viewer, apis);
-    Object.assign(instance.data, apis);
 
     instance.state = new ReactiveDict();
     instance.state.set('leftSidebar', Session.get('leftSidebar'));
@@ -152,98 +139,10 @@ Template.viewer.onCreated(() => {
         OHIF.viewerbase.updateMetaDataManager(study);
     });
 
-    const patientId = instance.data.studies[0].patientId;
-
-    // LT-382: Preventing HP to keep identifying studies in timepoints that might be removed
-    instance.data.studies.forEach(study => (delete study.timepointType));
-
-    // TODO: Consider combining the retrieval calls into one?
-    const timepointsPromise = timepointApi.retrieveTimepoints({ patientId });
-    timepointsPromise.then(() => {
-        const timepoints = timepointApi.all();
-
-        //  Set timepointType in studies to be used in hanging protocol engine
-        timepoints.forEach(timepoint => {
-            timepoint.studyInstanceUids.forEach(studyInstanceUid => {
-                const study = _.find(instance.data.studies, element => {
-                    return element.studyInstanceUid === studyInstanceUid;
-                });
-                if (!study) {
-                    return;
-                }
-
-                // @TODO: Maybe this should be a setCustomAttribute?
-                study.timepointType = timepoint.timepointType;
-            });
-        });
-
-        Session.set('TimepointsReady', true);
-
-        const timepointIds = timepoints.map(t => t.timepointId);
-
-        const measurementsPromise = measurementApi.retrieveMeasurements(patientId, timepointIds);
-        measurementsPromise.then(() => {
-            Session.set('MeasurementsReady', true);
-
-            measurementApi.syncMeasurementsAndToolData();
-        });
-    });
-
-    measurementApi.priorTimepointId = OHIF.viewer.data.currentTimepointId;
-
-    let firstMeasurementActivated = false;
-    instance.autorun(() => {
-        if (!Session.get('TimepointsReady') ||
-            !Session.get('MeasurementsReady') ||
-            !Session.get('ViewerReady') ||
-            firstMeasurementActivated) {
-            return;
-        }
-
-        // Find and activate the first measurement by Lesion Number
-        // NOTE: This is inefficient, we should be using a hanging protocol
-        // to hang the first measurement's imageId immediately, rather
-        // than changing images after initial loading...
-        const config = OHIF.measurements.MeasurementApi.getConfiguration();
-        const tools = config.measurementTools[0].childTools;
-        const firstTool = tools[Object.keys(tools)[0]];
-        const measurementTypeId = firstTool.id;
-
-        const collection = measurementApi.tools[measurementTypeId];
-        const sorting = {
-            sort: {
-                measurementNumber: -1
-            }
-        };
-
-        const data = collection.find({}, sorting).fetch();
-
-        let timepoints = [OHIF.viewer.data.currentTimepointId];
-
-        // TODO: Clean this up, it's probably an inefficient way to get what we need
-        const groupObject = _.groupBy(data, m => m.measurementNumber);
-
-        // Reformat the data
-        const rows = Object.keys(groupObject).map(key => ({
-            measurementTypeId: measurementTypeId,
-            measurementNumber: key,
-            entries: groupObject[key]
-        }));
-
-        const rowItem = rows[0];
-
-        // Activate the first lesion
-        if (rowItem) {
-            OHIF.measurements.jumpToRowItem(rowItem, timepoints);
-        }
-
-        firstMeasurementActivated = true;
-    });
-
-    instance.measurementModifiedHandler = _.throttle((event, instance) => {
-        OHIF.measurements.MeasurementHandlers.onModified(event, instance);
-    }, 300);
-
+    // Call Viewer plugins onCreated functions
+    if(typeof OHIF.viewer.measurementTable.onCreated === 'function') {
+        OHIF.viewer.measurementTable.onCreated(instance);
+    }
 });
 
 Template.viewer.onRendered(function() {
@@ -260,42 +159,40 @@ Template.viewer.onRendered(function() {
         }
     });
 
-});
-
-Template.viewer.events({
-    'click .js-toggle-studies'() {
-        const instance = Template.instance();
-        const current = instance.state.get('leftSidebar');
-        instance.state.set('leftSidebar', !current);
-    },
-
-    'click .js-toggle-protocol-editor'() {
-        const instance = Template.instance();
-        const current = instance.state.get('rightSidebar');
-        instance.data.state.set('rightSidebar', !current);
-    },
-
-    'cornerstonetoolsmeasurementadded .imageViewerViewport'(event, instance) {
-        const originalEvent = event.originalEvent;
-        OHIF.measurements.MeasurementHandlers.onAdded(originalEvent, instance);
-    },
-
-    'cornerstonetoolsmeasurementmodified .imageViewerViewport'(event, instance) {
-        const originalEvent = event.originalEvent;
-        instance.measurementModifiedHandler(originalEvent, instance);
-    },
-
-    'cornerstonemeasurementremoved .imageViewerViewport'(event, instance) {
-        const originalEvent = event.originalEvent;
-        OHIF.measurements.MeasurementHandlers.onRemoved(originalEvent, instance);
+    // Call Viewer plugins onRendered functions 
+    if(typeof OHIF.viewer.measurementTable.onRendered === 'function') {
+        OHIF.viewer.measurementTable.onRendered();
     }
+
 });
 
-Template.viewer.onDestroyed(() => {
-    Session.set('TimepointsReady', false);
-    Session.set('MeasurementsReady', false);
+Template.viewer.events(function() {
+    const events = {
+        'click .js-toggle-studies'() {
+            const instance = Template.instance();
+            const current = instance.state.get('leftSidebar');
+            instance.state.set('leftSidebar', !current);
+        },
 
-    OHIF.viewer.stackImagePositionOffsetSynchronizer.deactivate();
+        'click .js-toggle-protocol-editor'() {
+            const instance = Template.instance();
+            const current = instance.state.get('rightSidebar');
+            instance.data.state.set('rightSidebar', !current);
+        }
+    };
+
+    // Appending possible events coming from Viewer plugins
+    if (typeof OHIF.viewer.measurementTable.events === 'function') {
+        Object.assign(events, OHIF.viewer.measurementTable.events());
+    }
+
+    return events;
+});
+
+Template.viewer.onDestroyed(function() {
+    if(typeof OHIF.viewer.measurementTable.onDestroyed === 'function') {
+        OHIF.viewer.measurementTable.onDestroyed();
+    }
 });
 
 Template.viewer.helpers({
