+++ Packages/ohif-measurement-table/client/configuration/configuration.js
@@ -3,23 +3,25 @@ import { OHIF } from 'meteor/ohif:core';
 import { measurementTools } from './measurementTools';
 import { retrieveMeasurements, storeMeasurements, retrieveTimepoints, storeTimepoints, removeTimepoint, updateTimepoint, disassociateStudy } from './dataExchange';
 
-OHIF.measurements.MeasurementApi.setConfiguration({
-    measurementTools,
-    dataExchange: {
-        retrieve: retrieveMeasurements,
-        store: storeMeasurements
-    },
-    dataValidation: {
-        validation: () => {}
-    }
-});
+export const configureApis = () => {
+    OHIF.measurements.MeasurementApi.setConfiguration({
+        measurementTools,
+        dataExchange: {
+            retrieve: retrieveMeasurements,
+            store: storeMeasurements
+        },
+        dataValidation: {
+            validation: () => {}
+        }
+    });
 
-OHIF.measurements.TimepointApi.setConfiguration({
-    dataExchange: {
-        retrieve: retrieveTimepoints,
-        store: storeTimepoints,
-        remove: removeTimepoint,
-        update: updateTimepoint,
-        disassociate: disassociateStudy
-    }
-});
+    OHIF.measurements.TimepointApi.setConfiguration({
+        dataExchange: {
+            retrieve: retrieveTimepoints,
+            store: storeTimepoints,
+            remove: removeTimepoint,
+            update: updateTimepoint,
+            disassociate: disassociateStudy
+        }
+    });
+};
