--- Packages/ohif-measurement-table/client/configuration/index.js
@@ -1,3 +1,0 @@
-import './measurementTools.js';
-import './dataExchange.js';
-import './configuration.js';
